==> 98. PySpark_Project/1_CaseStudyOnLargeDataSetProcessing/4_case_study.ipynb <==
# Loading the DataSet

from pyspark.sql import SparkSession
# Set log level to ERROR (will suppress warnings)


spark = SparkSession.builder.appName('CustomerTransactionCaseStudy').getOrCreate()
spark.sparkContext.setLogLevel("ERROR")


df_large = spark.read.csv('./source/customer_transactions.csv', header=True, inferSchema=True)
df_large.show(10)

# Data Exploration
- check the schema and inspect some basic statistics to ensure the data is consistent

df_large.printSchema()
df_large.select('customer_id', 'purchase_date', 'amount_spent').describe().show()

#Check data distribution
df_large.groupBy('location').count().show()

# Data Cleaning
- Remove duplicates
- Drop rows where amount_spent is missing


df_clean = df_large.dropDuplicates()

df_clean = df_clean.filter(df_clean.amount_spent.isNotNull())
df_clean.printSchema()

from pyspark.sql.functions import col, to_date
df_clean = df_clean.withColumn('purchase_date', to_date(col('purchase_date'), 'yyyy-MM-dd'))
df_clean.show(10)

df_clean.printSchema()

# Data Aggregation
- Perform some aggregations to compute key metrics.

import time
from pyspark.sql.functions import sum

start_time = time.time()
customer_summary = df_clean.groupBy('customer_id').agg(
    sum('amount_spent').alias('total_spent')
)
customer_summary.show(5)
end_time = time.time()
print(f'Execution Time : {end_time - start_time:.2f} seconds')

### Challenge: Large Aggregation Operations
* Aggregations can cause shuffling of the data, which is expensive when working with large datasets
* We can use partitioning or bucketing to minimize the shuffle during the aggregation process

import time
from pyspark.sql.functions import sum

# repartitioning the data into 100 partitions in a way that all rows with the same value for customer_id are likely to end up in the same partition
df_clean_repartitioned = df_clean.repartition(100, 'customer_id') 

start_time = time.time()
customer_summary = df_clean_repartitioned.groupBy('customer_id').agg(
    sum('amount_spent').alias('total_spent')
)
customer_summary.show(5)
end_time = time.time()
print(f'Execution Time : {end_time - start_time:.2f} seconds')

# Identifying Top Customers

top_customers = customer_summary.orderBy('total_spent', ascending=False).limit(10)
top_customers.show(5)

# Joining with Product and Demographic Data

import time
product_info = spark.read.csv('./source/product_info.csv', header=True, inferSchema=True)
customer_info = spark.read.csv('./source/customer_info.csv', header=True, inferSchema=True)

start_time = time.time()
df_joined = df_clean.join(product_info, "product_id")
df_full = df_joined.join(customer_info, "customer_id")

df_full.show(10)
end_time = time.time()
print(f"Execution time : {end_time - start_time:.2f} seconds")

# Optimizations during Join
- Joins can cause shuffling and are often the slowest operation when processing large datasets. We can optimize joins by:

  - Broadcasting the smaller `product_info` dataframe


import time
from pyspark.sql.functions import broadcast

product_info = spark.read.csv('./source/product_info.csv', header=True, inferSchema=True)
customer_info = spark.read.csv('./source/customer_info.csv', header=True, inferSchema=True)

start_time = time.time()
df_joined = df_clean.join(broadcast(product_info), "product_id")
df_full = df_joined.join(customer_info, "customer_id")

df_full.show(10)
end_time = time.time()
print(f"Execution time : {end_time - start_time :.2f} seconds")

# Writing Results
- After performing all the necessary transformations, write the results back to storage.
- It's critical to optimize the write process to avoid generating too many small files

df_full.coalesce(1).write.mode('overwrite').parquet('./target/fully_joined_dataset.parquet')

# Monitoring and Profiling

#### Understanding Plan
Parsed Logical Plan - shows the SQL-like plan before optimization; Displays operations like JOIN, FILTER, PROJECT, AGGREGATE

Analyzed Logical Plan - checks column names and types; Ensures column `customer_id` is valid and `amount_spent` is `double`, etc

Optimized Logical Plan - applies optimizations like: Predicate pushdown (filters only), Broadcast join (small tables sent to all workers), Column pruning (removes unused columns)

Physical Plan - Spark physically executes the query. Uses: BroadcastHashJoin, SortMergeJoin, Exchange operations
    - Displays `partitioning strategy` and `file scan details`

df_full.explain(True)

# SUMMARY

### Challenges Faced:
Data Shuffling: Shuffling happens during operations like groupBy and join. This can cause performance bottlenecks. Optimization: Partitioning the data appropriately and using broadcast joins for smaller datasets.

Memory Usage: Large aggregations can overwhelm memory. Optimization: Use partitioning and caching appropriately to ensure data fits within the memory.

File Handling: Writing large datasets may result in too many small files. Optimization: Use .coalesce() to reduce the number of output files.



### In this case study, we:

Loaded and cleaned large datasets.

Aggregated the data efficiently using partitioning.

Performed joins using broadcast strategies.

Written optimized results back to storage in a performant way.
